pyproject.toml
[tool.poetry]
name = "diploma-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.idea
    | \.git
    | \.venv
  )/
  | manage.py
)
'''

[tool.isort]
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


